find min/max u/v/w norm in calc directly
... w/o using cone2mesh, and use that the find cmin & cmax
    which now correspond to "absolute" vector norm.

    This make cone/helpers.js obsolete.
